Extract loading examples and the CSS of the app to separate files, to keep the actual example code smaller.
